refactor: split AppStoreTests into per-concern suites with shared fixtures
