+++ agtermCore/Tests/agtermCoreTests/AppStoreTestFixtures.swift
@@ -1,0 +1,15 @@
+import Foundation
+@testable import agtermCore
+
+/// A store backed by a throwaway temp directory so mutation-time saves never
+/// touch the real Application Support path. PersistenceStore creates the
+/// directory lazily on first write.
+@MainActor func makeStore() -> AppStore {
+    let dir = FileManager.default.temporaryDirectory.appendingPathComponent("agterm-tests-\(UUID().uuidString)")
+    return AppStore(persistence: PersistenceStore(directory: dir))
+}
+
+final class SpySurface: TerminalSurface {
+    var teardownCount = 0
+    func teardown() { teardownCount += 1 }
+}
