+++ docs/plans/20260701-file-size-limits-refactor.md
@@ -183,11 +183,11 @@ Read `.claude/rules/ui-tests.md` first. CRITICAL: nothing in the normal toolchai
 
 Swift Testing; sibling convention is `@MainActor struct` with no `@Suite` attribute. Capture the FULL-MODULE `swift test` executed-test baseline total BEFORE step 1 (this file alone contributes 178 `@Test` funcs).
 
-- [ ] create `AppStoreTestFixtures.swift` (imports Foundation + `@testable import agtermCore`): move `makeStore` verbatim (doc comment included) as internal top-level `@MainActor func makeStore() -> AppStore`, and `SpySurface` verbatim with `private` dropped — do NOT add an explicit `@MainActor` annotation to SpySurface (it is MainActor-inferred via its `TerminalSurface` conformance, declaration-based, so the verbatim move is isolation-preserving). In AppStoreTests.swift delete the originals and replace all 143 `Self.makeStore()` with `makeStore()` (pre-verified: exactly 143, all zero-arg, no other makeStore in the module). Verify: `rg -c 'Self.makeStore'` = 0; full-module executed-test total unchanged vs baseline; `make lint`; `make build`.
-- [ ] create `AppStoreNavigationTests.swift`: move the navigation regions (orig 1394–1405 + 1743–2077) byte-verbatim into `@MainActor struct AppStoreNavigationTests`; static `makeNavTree` travels with all 19 `Self.makeNavTree()` sites. Collapse each excision seam in the residual to ONE blank line (`vertical_whitespace` fails `--strict` on doubles). Verify: full-module total unchanged; lint.
-- [ ] create `AppStoreOrganizationTests.swift`: move orig 1071–1091, 1152–1260, 1262–1392 stitched with a single blank line to 1407–1482 (the seam left by the navigation lift), then 1484–1741; `makeReorderTree` + `makeWorkspaceReorderTree` statics travel with all 17 call sites. Collapse seams. Verify: total unchanged; lint.
-- [ ] create `AppStorePaneTests.swift`: move orig 419–608, 741–831, 833–943 byte-verbatim; add three `// MARK: -` section comments (split panes / overlay / scratch). Collapse seams. Verify: total unchanged; lint. (Name pairs with Task 10's `AppStore+Panes.swift` — do not rename unilaterally.)
-- [ ] final verification: `wc -l` the five files (~745 residual / ~405 / ~610 / ~360 / ~18 fixtures); `rg -c '@Test '` across the five sums to 178; `make build` + `cd agtermCore && swift test` (full-module total = baseline) + `make lint`.
+- [x] create `AppStoreTestFixtures.swift` (imports Foundation + `@testable import agtermCore`): move `makeStore` verbatim (doc comment included) as internal top-level `@MainActor func makeStore() -> AppStore`, and `SpySurface` verbatim with `private` dropped — do NOT add an explicit `@MainActor` annotation to SpySurface (it is MainActor-inferred via its `TerminalSurface` conformance, declaration-based, so the verbatim move is isolation-preserving). In AppStoreTests.swift delete the originals and replace all 143 `Self.makeStore()` with `makeStore()` (pre-verified: exactly 143, all zero-arg, no other makeStore in the module). Verify: `rg -c 'Self.makeStore'` = 0; full-module executed-test total unchanged vs baseline; `make lint`; `make build`.
+- [x] create `AppStoreNavigationTests.swift`: move the navigation regions (orig 1394–1405 + 1743–2077) byte-verbatim into `@MainActor struct AppStoreNavigationTests`; static `makeNavTree` travels with all 19 `Self.makeNavTree()` sites. Collapse each excision seam in the residual to ONE blank line (`vertical_whitespace` fails `--strict` on doubles). Verify: full-module total unchanged; lint.
+- [x] create `AppStoreOrganizationTests.swift`: move orig 1071–1091, 1152–1260, 1262–1392 stitched with a single blank line to 1407–1482 (the seam left by the navigation lift), then 1484–1741; `makeReorderTree` + `makeWorkspaceReorderTree` statics travel with all 17 call sites. Collapse seams. Verify: total unchanged; lint.
+- [x] create `AppStorePaneTests.swift`: move orig 419–608, 741–831, 833–943 byte-verbatim; add three `// MARK: -` section comments (split panes / overlay / scratch). Collapse seams. Verify: total unchanged; lint. (Name pairs with Task 10's `AppStore+Panes.swift` — do not rename unilaterally.)
+- [x] final verification: `wc -l` the five files (~745 residual / ~405 / ~610 / ~360 / ~18 fixtures); `rg -c '@Test '` across the five sums to 178; `make build` + `cd agtermCore && swift test` (full-module total = baseline) + `make lint`. (actuals: 741 residual / 407 pane / 606 org / 355 nav / 15 fixtures; @Test sum = 178 (66+26+57+29+0); make build OK; swift test 894 tests in 39 suites = baseline; make lint clean.)
 
 ### Task 7: Commands.swift — split by command family
